+++ CHANGES.rst
@@ -9,6 +9,7 @@ Version 3.X.X (2019-09-XX)
 - require ``urlquote>=1.1.0`` (where ``urlquote.quoting`` was introduced)
 - ``MetaPartition.parse_input_to_metapartition`` accepts dicts and list of tuples equivalents as ``obj`` input
 - Improve performance for some cases where predicates are used with the `in` operator.
+- Correctly preserve :class:`~kartothek.core.index.ExplicitSecondaryIndex` dtype when index is empty
 
 
 Version 3.4.0 (2019-09-17)